feat: configure react and slim extention module
