+++ package.json
@@ -29,11 +29,15 @@
 		"react": "^18.0.0",
 		"react-dom": "^18.0.0",
 		"style-loader": "^3.0.0",
+		"styled-components": "^5.3.9",
 		"uuid": "^9.0.0"
 	},
 	"devDependencies": {
 		"@types/jest": "^29.4.0",
 		"@types/jest-environment-puppeteer": "^5.0.0",
+		"@types/react-dom": "^18.0.11",
+		"@types/styled-components": "^5.1.26",
+		"@types/webextension-polyfill": "^0.10.0",
 		"eslint": "~8.6.0",
 		"html-webpack-plugin": "^5.5.0",
 		"jest": "^29.5.0",
