+++ src/index.html
@@ -6,7 +6,8 @@
 		<title>Coolighter</title>
 	</head>
 	<body style="margin: 0; height: 100%; display: flex">
-		<div id="left-hltr" style="min-height: 100%; width: 50%"></div>
-		<div id="right-hltr" style="min-height: 100%; width: 50%"></div>
+		<!-- <div id="left-hltr" style="min-height: 100%; width: 50%"></div>
+		<div id="right-hltr" style="min-height: 100%; width: 50%"></div> -->
+		<div id = "root"></div>
 	</body>
 </html>
