+++ src/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./Highlight/Components/App";
+
+const root = ReactDOM.createRoot(document.getElementById("root")!);
+
+root.render(<App  author="John Doe"/>);
