+++ webpack.config.js
@@ -9,7 +9,7 @@ module.exports = (env) => {
 	return merge(
 		{
 			entry: {
-				app: "./src/app.ts",
+				app: "./src/index.tsx",
 			},
 			resolve: {
 				extensions: [".ts", ".tsx", ".js"],
